fix(productos): monedas fijas COP/USD/UE + precio double; ui selector

+++ backend/src/main/java/com/diegoabarajas/pruebatecnica/producto/ProductoPrecioResponse.java
@@ -1,10 +1,8 @@
 package com.diegoabarajas.pruebatecnica.producto;
 
-import java.math.BigDecimal;
-
 public record ProductoPrecioResponse(
 		String moneda,
-		BigDecimal precio
+		Double precio
 ) {
 	public static ProductoPrecioResponse fromEntity(ProductoPrecio p) {
 		return new ProductoPrecioResponse(p.getMoneda(), p.getPrecio());
